# swagger_json.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功，请查收验证邮件",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "注册成功，请查看邮箱并点击验证链接完成注册",
                    "email": "user@example.com"
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "409": {
            "description": "邮箱或用户名已存在"
          }
        },
        "summary": "用户注册",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "账号或密码错误"
          }
        },
        "summary": "用户登录",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "无效的刷新Token"
          }
        },
        "summary": "刷新Token",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "登出成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "用户登出",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取用户信息成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户信息",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/send-verification": {
      "post": {
        "operationId": "AuthController_sendVerification",
        "parameters": [],
        "responses": {
          "200": {
            "description": "验证邮件已发送",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendVerificationApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或发送过于频繁"
          }
        },
        "summary": "发送邮箱验证码",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "邮箱验证成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "验证码无效或已过期"
          }
        },
        "summary": "验证邮箱",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "responses": {
          "200": {
            "description": "验证邮件已重新发送",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendVerificationApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或发送过于频繁"
          }
        },
        "summary": "重发验证码",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码重置验证码已发送",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "该邮箱未注册或账号已禁用"
          }
        },
        "summary": "忘记密码",
        "tags": [
          "认证"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码重置成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "验证码无效或已过期"
          }
        },
        "summary": "重置密码",
        "tags": [
          "认证"
        ]
      }
    },
...
  },
  "info": {
    "title": "CloudCAD API",
    "description": "图纸管理平台API文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "example": "user@example.com",
            "format": "email"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "username",
            "minLength": 3,
            "maxLength": 20,
            "pattern": "^[a-zA-Z0-9_]+$"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "password123",
            "minLength": 8,
            "maxLength": 50
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "用户昵称",
            "maxLength": 50
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "description": "邮箱或用户名",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "Password123!"
          }
        },
        "required": [
          "account",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "username"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL",
            "example": "https://example.com/avatar.jpg"
          },
          "role": {
            "type": "object",
            "description": "用户角色",
            "properties": {
              "id": {
                "type": "string",
                "example": "clxxxxxxx"
              },
              "name": {
                "type": "string",
                "enum": [
                  "ADMIN",
                  "USER_MANAGER",
                  "FONT_MANAGER",
                  "USER"
                ],
                "example": "USER"
              },
              "description": {
                "type": "string",
                "example": "普通用户，基础权限",
                "nullable": true
              },
              "isSystem": {
                "type": "boolean",
                "example": true
              },
              "permissions": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "permission": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "status": {
            "type": "string",
            "description": "用户状态",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ],
            "example": "ACTIVE"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "role",
          "status"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "访问Token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "刷新Token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "description": "用户信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "AuthApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应状态码",
            "example": "SUCCESS",
            "enum": [
              "SUCCESS",
              "ERROR"
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "description": "认证响应数据",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthResponseDto"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间戳",
            "example": "2025-12-12T03:34:55.801Z"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新Token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "UserRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "角色 ID"
          },
          "name": {
            "type": "string",
            "description": "角色名称"
          },
          "description": {
            "type": "string",
            "description": "角色描述"
          },
          "isSystem": {
            "type": "boolean",
            "description": "是否为系统角色"
          }
        },
        "required": [
          "id",
          "name",
          "isSystem"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户 ID"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像 URL"
          },
          "status": {
            "type": "string",
            "description": "用户状态",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          },
          "role": {
            "description": "用户角色",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRoleDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "status",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "SendVerificationResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "操作结果",
            "example": "验证邮件已发送"
          }
        },
        "required": [
          "message"
        ]
      },
      "SendVerificationApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/SendVerificationResponseDto"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SendVerificationResponseDto"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "6位数字验证码",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "VerifyEmailResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "操作结果",
            "example": "邮箱验证成功"
          }
        },
        "required": [
          "message"
        ]
      },
      "VerifyEmailApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/VerifyEmailResponseDto"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/VerifyEmailResponseDto"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "消息"
          }
        },
        "required": [
          "message"
        ]
      },
      "ForgotPasswordApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应状态码",
            "example": "SUCCESS",
            "enum": [
              "SUCCESS",
              "ERROR"
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "description": "响应数据",
            "allOf": [
              {
                "$ref": "#/components/schemas/ForgotPasswordResponseDto"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间戳",
            "example": "2025-12-12T03:34:55.801Z"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "邮箱地址",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "验证码",
            "example": "123456"
          },
          "newPassword": {
            "type": "string",
            "description": "新密码",
            "example": "NewPassword123!"
          },
          "confirmPassword": {
            "type": "string",
            "description": "确认新密码",
            "example": "NewPassword123!"
          }
        },
        "required": [
          "email",
          "code",
          "newPassword",
          "confirmPassword"
        ]
      },
      "ResetPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "消息"
          }
        },
        "required": [
          "message"
        ]
      },
      "ResetPasswordApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "响应状态码",
            "example": "SUCCESS",
            "enum": [
              "SUCCESS",
              "ERROR"
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "description": "响应数据",
            "allOf": [
              {
                "$ref": "#/components/schemas/ResetPasswordResponseDto"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间戳",
            "example": "2025-12-12T03:34:55.801Z"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
...
    }
  }
}